internal/contour: remove CDS translator code and tests
